Beta 8
UPDATE TO 1.19: Im waiting the optifine sources of 1.19 for update the client so for now im on 1.18.2. Dont ask for remove the future chat reporting in 1.19.1.

MAJOR: Added the BlockInfoTooltip Plugin (a plugin like wthit), it is not a hud plugin because his position is anchored on the screen (at the top on the center). Can change in the future
MAJOR: The client support capes with a width and a height of 32, 64, 128, 256, 512, 1024 and 2048px!!! the height should be the half of the width

Renamed CommonLezardComponents to CommonLezardVariables (for more accuracy)
Added a common translation for "Second Color"
Fixed a bug that call the start method of the client when you change/reload the resourcepacks